Initialize Jax Distributed System first
