+++ MaxText/checkpointing.py
@@ -18,38 +18,13 @@
 
 from etils import epath
 import jax
-import portpicker
-from jax.experimental import multihost_utils
 from orbax import checkpoint
 from orbax.checkpoint.checkpoint_manager import CheckpointManager, CheckpointManagerOptions, Checkpointer, AsyncCheckpointer
 from orbax.checkpoint import type_handlers
-import socket
 
 import max_logging
 
 from flax.training import train_state
-
-def _multislice_distribute_initialize():
-  """Calls jax.distribute.initialize() with appropriate multislice arguments."""
-
-  def gen_local_ip():
-    hostname = socket.gethostname()
-    return socket.gethostbyname(hostname)
-
-  def gen_local_ip_nums():
-    return [int(num) for num in gen_local_ip().split(':')[-1].split('.')]
-
-  def get_coordinator_ip():
-    local_ip_nums = jax.numpy.array(gen_local_ip_nums())
-    coordinator_ip_nums = multihost_utils.broadcast_one_to_all(local_ip_nums)
-    coordinator_ip_strings = [str(num) for num in list(coordinator_ip_nums)]
-    return '.'.join(coordinator_ip_strings)
-
-  port = multihost_utils.broadcast_one_to_all(jax.numpy.array(portpicker.pick_unused_port()))
-  coordinator_address = get_coordinator_ip() + ':' + str(port)
-  jax.distributed.initialize(coordinator_address=coordinator_address,
-                             num_processes=jax.process_count(),
-                             process_id=jax.process_index())
 
 def create_orbax_checkpoint_manager(
     checkpoint_dir: str,
@@ -64,7 +39,6 @@ def create_orbax_checkpoint_manager(
   max_logging.log("Creating checkpoint manager...")
   p = epath.Path(checkpoint_dir)
   if use_async:
-    _multislice_distribute_initialize()
     checkpointer = AsyncCheckpointer(checkpoint.PyTreeCheckpointHandler())
   else:
     checkpointer = Checkpointer(checkpoint.PyTreeCheckpointHandler())
